LayoutTests/canvas/philip/tests/2d.state.saverestore.transformation-expected.txt

< [index] >
2d.state.saverestore.transformation
save()/restore() affects the current transformation matrix
References: 2d.state.transformation
Actual output:
Expected output:

Passed
